requirements.txt
discord.py[voice]==1.3.2

